Separate out transparent touch view into the extensions folder
